Move merge_config to shared utils
